cleanup cmake a bit more

+++ bbs/dirlist.cpp
@@ -178,7 +178,7 @@ void dirlist(int mode) {
         is = false;
       }
       if (isdigit(ss.front())) {
-        for (size_t i3 = 0; i3 < a()->directories.size(); i3++) {
+        for (uint16_t i3 = 0; i3 < a()->directories.size(); i3++) {
           if (ss == a()->udir[i3].keys) {
             a()->set_current_user_dir_num(i3);
             os = a()->current_user_dir().subnum;
